move test to config.rs
